split The algaeByeBye into two subsystems one for each motor
