+++ src/main/java/frc/robot/subsystems/AlgaeByeBye/AlgaeByeByeInputs.java
@@ -1,0 +1,23 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.subsystems.AlgaeByeBye;
+
+import org.littletonrobotics.junction.LogTable;
+import org.littletonrobotics.junction.inputs.LoggableInputs;
+
+/** Add your docs here. */
+public class AlgaeByeByeInputs implements LoggableInputs {
+
+  @Override
+  public void toLog(LogTable table) {
+
+  }
+
+  @Override
+  public void fromLog(LogTable table) {
+
+  }
+ 
+}
