+++ src/main/java/frc/robot/subsystems/AlgaeByeBye/MockAlgaeByeByeIO.java
@@ -2,25 +2,17 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.AlgaeRemover;
+package frc.robot.subsystems.AlgaeByeBye;
+
 
 /** Add your docs here. */
-public class MockAlgaeRemoverIO implements AlgaeRemoverIO {
+public class MockAlgaeByeByeIO implements AlgaeByeByeIO {
   @Override
   public void setRemoverSpeed(double speed) {}
-
-  @Override
-  public void setTiltMotorSpeed(double tiltSpeed) {}
 
   @Override
   public void stopRemoverMotors() {}
 
   @Override
-  public void stopTiltMotors() {}
-
-  @Override
-  public void resetTiltEncoder() {}
-
-  @Override
-  public void updateInputs(AlgaeRemoverInputs inputs) {}
+  public void updateInputs(AlgaeByeByeInputs inputs) {}
 }
