+++ src/main/java/frc/robot/subsystems/AlgaeByeByeTilt/AlgaeByeByeTiltInputs.java
@@ -2,14 +2,14 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.AlgaeRemover;
+package frc.robot.subsystems.AlgaeByeByeTilt;
 
 import org.littletonrobotics.junction.LogTable;
 import org.littletonrobotics.junction.inputs.LoggableInputs;
 
 /** Add your docs here. */
-public class AlgaeRemoverInputs implements LoggableInputs {
-  public double tiltMotorEncoderPosition = 0;
+public class AlgaeByeByeTiltInputs implements LoggableInputs{
+ public double tiltMotorEncoderPosition = 0;
   public boolean backLimitSwitchState;
   public boolean forwardLimitSwitchState;
 
@@ -24,4 +24,5 @@ public class AlgaeRemoverInputs implements LoggableInputs {
     backLimitSwitchState = table.get("backLimitSwitchState", backLimitSwitchState);
     forwardLimitSwitchState = table.get("forwardLimitSwitchState", forwardLimitSwitchState);
   }
+
 }
