+++ src/main/java/frc/robot/subsystems/AlgaeByeByeTilt/MockAlgaeByeByeTIltIO.java
@@ -1,0 +1,21 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+package frc.robot.subsystems.AlgaeByeByeTilt;
+
+/** Add your docs here. */
+public class MockAlgaeByeByeTIltIO implements AlgaeByeByeTiltIO{
+
+@Override
+public void updateInputs(AlgaeByeByeTiltInputs inputs) {}
+
+@Override
+public void setTiltSpeed(double speed) {}
+
+@Override
+public void stopTiltMotors() {}
+
+@Override
+public void resetTiltEncoder() {}
+}
